Add skeleton QALSource class into the project
QtOpengl was designed for solving the platform specific details. It was not
designed for a complete graphics engine. However, the starting point is
different for a project, like QtOpenAL since openal already solves the platform
specific details. Hence, it is already a very huge design decision difference.

The QtOpenAL API users would not probably like to convert qt types and so on
manually. It would be better to implement a full wrapper since there is not much
need for solving platform specific details apart from having conditional include
for the relevant platform specific openal files.

Summary: The purpose is to make classes and method for everything that a 3D
audio Qt programmer would like to achieve. The API users could get rid of the
low-level openal usage.