Do not fail in the reset method, if the alcCloseDevice fails
If the close fails, the context may already be gone. If it returns false,
the application would assume it did not reset and is still valid.

Also the pointers need to be set to 0 after destroying and/or closing them.

There is a need for a separate and explicite error check since the context
destroy method returns void. ALC_INVALID_CONTEXT is the only possible error
according to the 1.1 specification.