Do not use QMutexLocker, just a private class member QMutex for locking
We need to construct a mutex anyway since the QMutexLocker does not have a
constructor which would do that automatically for us. Also, it is better to
unlock immediately after the mpg123_init call in order to lock as less code as
needed. This way, the construction and deconstruction of a mutex and locker does
not take effect on the private class' constructor and destructor's speed.